split `TestSuitesSourceController` to internal and public controllers
